+++ Crossy Road/Game/State/state_playing.cpp
@@ -72,8 +72,8 @@ std::vector<SaveInf::RoadInf> state_playing::randomSaveInf(unsigned lv)
     for (size_t i = 0; i < SAVE_LANE; ++i) {
         direction   = -1 + r.int_in_range(0, 1) * 2;//left or right
         vehicleType = r.int_in_range(0, maxVehicleType);//random base on how many type read from Config/
-        vehicleNum = lv < 5 ? (r.int_in_range(1, lv)) : (lv < 10 ? r.int_in_range(2, 5) :r.int_in_range(3,5));
-        speed = lv>2?(r.double_in_range(-10.f, 10.f) + 10.f * (float)lv):r.double_in_range(10.f,20.f);
+        vehicleNum = lv < 5 ? (r.int_in_range(1, lv)) : (lv < 10 ? r.int_in_range(2, 5) : r.int_in_range(3, 5));
+        speed = lv > 2 ? (r.double_in_range(-10.f, 10.f) + 10 * lv) : r.double_in_range(10.f, 20.f);
 
         lane.emplace_back(SaveInf::RoadInf({ vehicleType, vehicleNum, direction, speed }));
     }
